+++ Rhombus-Editor/src/Panels/TilesetPanel.cpp
@@ -28,11 +28,11 @@ namespace rhombus
 			{
 				ImGuiStyle& style = ImGui::GetStyle();
 				auto& colors = style.Colors;
-				for (int row = 0; row < m_Tileset->GetRowCount(); row++)
+				for (uint32_t row = 0; row < m_Tileset->GetRowCount(); row++)
 				{
 					ImGui::PushStyleVar(ImGuiStyleVar_CellPadding, ImVec2(style.CellPadding.x, 0.0f));
 					ImGui::TableNextRow();
-					for (int column = 0; column < m_Tileset->GetColumnCount(); column++)
+					for (uint32_t column = 0; column < m_Tileset->GetColumnCount(); column++)
 					{
 						ImGui::TableSetColumnIndex(column);
 						int tileIndex = column + row * m_Tileset->GetColumnCount();
@@ -46,7 +46,7 @@ namespace rhombus
 							ImGui::PushStyleColor(ImGuiCol_Button, style.Colors[ImGuiCol_ButtonActive]);
 						}
 
-						if (ImGui::ImageButton((ImTextureID)tileTexture->GetTexture()->GetRendererID(), { thumbnailSize, thumbnailSize }, ImVec2(tileTexture->GetTexCoords()[3].x, tileTexture->GetTexCoords()[3].y), ImVec2(tileTexture->GetTexCoords()[1].x, tileTexture->GetTexCoords()[1].y)))
+						if (ImGui::ImageButton((ImTextureID)(intptr_t)tileTexture->GetTexture()->GetRendererID(), { thumbnailSize, thumbnailSize }, ImVec2(tileTexture->GetTexCoords()[3].x, tileTexture->GetTexCoords()[3].y), ImVec2(tileTexture->GetTexCoords()[1].x, tileTexture->GetTexCoords()[1].y)))
 						{
 							if (tileIndex != m_iSelectedTileIndex)
 							{
