+++ Rhombus/src/Rhombus/ECS/Components/SpriteRendererComponent.cpp
@@ -22,11 +22,11 @@ namespace rhombus
 			Vec2 spriteSize = Vec2(((float)m_texture->GetWidth() / (float)m_columns) - 2.0f * m_padding, ((float)m_texture->GetHeight() / (float)m_rows) - 2.0f * m_padding);
 			if (m_subtexture)
 			{
-				m_subtexture->UpdateTexCoords(Vec2(m_frame % m_columns, m_rows - 1.0f - m_frame / m_columns), spriteSize, m_padding);
+				m_subtexture->UpdateTexCoords(Vec2(m_frame % m_columns, m_rows - 1 - m_frame / m_columns), spriteSize, (float)m_padding);
 			}
 			else
 			{
-				m_subtexture = SubTexture2D::CreateFromCoords(m_texture, Vec2(m_frame % m_columns, m_rows - 1.0f - m_frame / m_columns), spriteSize, m_padding);
+				m_subtexture = SubTexture2D::CreateFromCoords(m_texture, Vec2(m_frame % m_columns, m_rows - 1 - m_frame / m_columns), spriteSize, (float)m_padding);
 			}
 		}
 	}
